+++ database/migrations/2019_02_18_062028_create_course_videos_table.php
@@ -21,7 +21,7 @@ class CreateCourseVideosTable extends Migration
             $table->string('duration', 50)->nullable();
             $table->text('image_name')->nullable();
             $table->text('video_tag')->nullable();
-            $table->unsignedInteger('uploader_id')->nullable();
+            $table->unsignedBigInteger('uploader_id')->nullable();
             $table->unsignedInteger('course_id')->nullable();
             $table->boolean('processed')->default(true);
             $table->timestamps();
